# api/openapi.local.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agent Lab API",
    "version": "0.1.0",
    "description": "Containerized web service platform for building and orchestrating agentic workflows."
  },
  "servers": [
    {
      "url": "http://localhost:8080"
    }
  ],
  "paths": {
...
    "/api/providers": {
      "get": {
        "summary": "List providers",
        "description": "Returns a paginated list of providers with optional filtering and sorting",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number (1-indexed)",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "description": "Results per page",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "search",
            "in": "query",
            "description": "Search query (matches name)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Comma-separated sort fields. Prefix with - for descending",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "name",
            "in": "query",
            "description": "Filter by provider name (contains)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of providers",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPageResult"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create provider",
        "description": "Validates and stores a new provider configuration",
        "tags": [
          "Providers"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProviderCommand"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Provider created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "400": {
            "description": "",
            "$ref": "#/components/responses/BadRequest"
          },
          "409": {
            "description": "",
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/providers/search": {
      "post": {
        "summary": "Search providers",
        "description": "Search providers with filters and pagination via POST body",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "query",
            "description": "Filter by provider name (contains)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Paginated search results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProviderPageResult"
                }
              }
            }
          },
          "400": {
            "description": "",
            "$ref": "#/components/responses/BadRequest"
          }
        }
      }
    },
    "/api/providers/{id}": {
      "get": {
        "summary": "Get provider by ID",
        "description": "Retrieves a single provider configuration",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Provider UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Provider configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "400": {
            "description": "",
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "",
            "$ref": "#/components/responses/NotFound"
          }
        }
      },
      "put": {
        "summary": "Update provider",
        "description": "Updates an existing provider configuration",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Provider UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProviderCommand"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Provider updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Provider"
                }
              }
            }
          },
          "400": {
            "description": "",
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "description": "",
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "description": "",
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "summary": "Delete provider",
        "description": "Removes a provider configuration",
        "tags": [
          "Providers"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Provider UUID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Provider deleted"
          },
          "404": {
            "description": "",
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateProviderCommand": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "description": "go-agents ProviderConfig as JSON"
          },
          "name": {
            "type": "string",
            "example": "ollama"
          }
        },
        "required": [
          "name",
          "config"
        ]
      },
...
      "PageRequest": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "description": "Page number (1-indexed)",
            "example": 1
          },
          "page_size": {
            "type": "integer",
            "description": "Results per page",
            "example": 20
          },
          "search": {
            "type": "string",
            "description": "Search query"
          },
          "sort": {
            "type": "string",
            "description": "Comma-separated sort fields. Prefix with - for descending. Example: name,-created_at"
          }
        }
      },
      "Provider": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "description": "go-agents ProviderConfig as JSON"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProviderPageResult": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "description": "Array of providers"
          },
          "page": {
            "type": "integer",
            "description": "Current page number"
          },
          "page_size": {
            "type": "integer",
            "description": "Results per page"
          },
          "total": {
            "type": "integer",
            "description": "Total number of results"
          },
          "total_pages": {
            "type": "integer",
            "description": "Total number of pages"
          }
        }
      },
...
      "UpdateProviderCommand": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "description": "go-agents ProviderConfig as JSON"
          },
          "name": {
            "type": "string"
          }
        },
        "required": [
          "name",
          "config"
        ]
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Invalid request",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            }
          }
        }
      },
      "Conflict": {
        "description": "Resource conflict (duplicate name)",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string",
                  "description": "Error message"
                }
              }
            }
          }
        }
      }
    }
  }
}